:art: Refactor greeting out of Agenda component
