+++ src/components/Agenda/Greeting/style.scss
@@ -1,0 +1,11 @@
+@import 'lib/style';
+
+.header {
+    display: block;
+    padding: 16px 0 12px 16px;
+  }
+  
+  .header {
+    @include title-text;
+    color: $blackColor;
+  }
