+++ src/components/Agenda/style.scss
@@ -23,16 +23,6 @@
   margin: 0 auto;
 }
 
-.header {
-  display: block;
-  padding: 16px 0 12px 16px;
-}
-
-.header {
-  @include title-text;
-  color: $blackColor;
-}
-
 .options {
   box-sizing: border-box;
   padding: 1em;
